Add windows includes to xmlIO.c
xmlIO.c calls read() and getcwd() which need io.h and direct.h
respectively when compiling on windows. Otherwise, a compiler error may
be raised saying that read() and getcwd() were used implicitly.

This was regressed recently, I'm guessing it was due to the changes to
win32config.h in commit 501e1b95