build: require automake-1.16.3 or later
We need
https://git.savannah.gnu.org/cgit/automake.git/commit/?id=e21d46f
to avoid an error like

    checking for PYTHON... no
    configure: error: Package requirements (python-3.1) were not met:

    No package 'python-3.1' found

With python-3.10.

Partially fixes #392.

Note that there is another issue with python-3.10, like:

    checking for python... no
    checking for python2... no
    checking for python3... no
    checking for python3.9... no
    ...
    configure: error: no suitable Python interpreter found

Fixing this one will need to require bumping automake version again, but
the automake version containing this fix is not released yet.  And, this
is rather minor because most environment has "python3" as a symlink to
"python3.10" or whatever the latest python3.x installed.