meson: Fix setup with ICU as sibling subproject
Meson wrapdb provides a wrap for ICU, so libxml2 and ICU could both be
built as subprojects of the same Meson parent project.  In this case, with
the icu option enabled, setup was failing with:

    subprojects/libxml2-2.13.5/meson.build:603:22: ERROR: Could not get an internal variable and no default provided for <InternalDependency dep228908115162702543524838879388991448872: True>

This is because we can't get a dependency variable from a subproject that
hasn't been built yet.  Fall back to assuming DEFS is empty, as it is on
my system.